app/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
greenlet==1.1.2
idna==3.3
numpy==1.22.1
pandas==1.4.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
urllib3==1.26.8
